src/test/java/com/leetcode/dp/medium/MaximalSquare_221Test.java
package com.leetcode.dp.medium;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class MaximalSquare_221Test {
  private final MaximalSquare_221 solution = new MaximalSquare_221();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(
          arguments(
              new char[][] {
                {'1', '0', '1', '0', '0'},
                {'1', '0', '1', '1', '1'},
                {'1', '1', '1', '1', '1'},
                {'1', '0', '0', '1', '0'}
              },
              4),
          arguments(new char[][] {{'0', '1'}, {'1', '0'}}, 1),
          arguments(new char[][] {{'0'}}, 0));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void maximalSquare(char[][] matrix, int expectedResult) {
    int actualResult = solution.maximalSquare(matrix);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
